Add unit tests for new page abstractions
